# server/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "This is a simple blog service to demonstrate Swagger."
  version: "1.0.0"
  title: "My Blog Service"
host: "localhost:8080"
basePath: "/my-blog-service/v1"
tags:
- name: "blogpost"
  description: "Access to blog posts"
schemes:
- "http"
paths:
  /blog:
    post:
      tags:
      - "blogpost"
      summary: "Add a blog post"
      operationId: "addBlogPost"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "New blog post"
        required: true
        schema:
          $ref: "#/definitions/BlogPost"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/BlogPost"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "Blogpost"
    put:
      tags:
      - "blogpost"
      summary: "Update an existing blog post"
      operationId: "updateBlogPost"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Blog post object to be updated"
        required: true
        schema:
          $ref: "#/definitions/BlogPost"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Blog post not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "Blogpost"
  /blog/{id}:
    get:
      tags:
      - "blogpost"
      summary: "Find blog post by ID"
      description: "Returns a single blog post"
      operationId: "getBlogPostById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of blog post to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/BlogPost"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Blog post not found"
      x-swagger-router-controller: "Blogpost"
    delete:
      tags:
      - "blogpost"
      summary: "Deletes a blog post"
      description: ""
      operationId: "deleteBlogPost"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the blog post to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Blog post not found"
      x-swagger-router-controller: "Blogpost"
definitions:
  BlogPost:
    type: "object"
    required:
    - "body"
    - "title"
    properties:
      id:
        type: "integer"
        format: "int64"
        example: 1
      title:
        type: "string"
        example: "My Blog Post"
      body:
        type: "string"
        example: "This is my blog about Swagger"
    example:
      id: 1
      title: "My Blog Post"
      body: "This is my blog about Swagger"